refactor(phase3a): hidden states 関連を library/hidden_states.py に分離
train_util.py の get_hidden_states / pool_workaround / get_hidden_states_sdxl
を library/hidden_states.py に移動。train_util.py は re-export shim として維持。
caching.py は循環依存が無いため _tu() 経由ではなく直接 import に変更。

train_util.py: 2,600 → 2,449 行
library/hidden_states.py: 173 行（新規）
