+++ library/train_util.py
@@ -1472,163 +1472,14 @@ def patch_accelerator_for_fp16_training(accelerator):
     accelerator.scaler._unscale_grads_ = _unscale_grads_replacer
 
 
-def get_hidden_states(args: argparse.Namespace, input_ids, tokenizer, text_encoder, weight_dtype=None):
-    # with no_token_padding, the length is not max length, return result immediately
-    if input_ids.size()[-1] != tokenizer.model_max_length:
-        return text_encoder(input_ids)[0]
-
-    # input_ids: b,n,77
-    b_size = input_ids.size()[0]
-    input_ids = input_ids.reshape((-1, tokenizer.model_max_length))  # batch_size*3, 77
-
-    if args.clip_skip is None:
-        encoder_hidden_states = text_encoder(input_ids)[0]
-    else:
-        enc_out = text_encoder(input_ids, output_hidden_states=True, return_dict=True)
-        encoder_hidden_states = enc_out["hidden_states"][-args.clip_skip]
-        encoder_hidden_states = text_encoder.text_model.final_layer_norm(encoder_hidden_states)
-
-    # bs*3, 77, 768 or 1024
-    encoder_hidden_states = encoder_hidden_states.reshape((b_size, -1, encoder_hidden_states.shape[-1]))
-
-    if args.max_token_length is not None:
-        if args.v2:
-            # v2: <BOS>...<EOS> <PAD> ... の三連を <BOS>...<EOS> <PAD> ... へ戻す　正直この実装でいいのかわからん
-            states_list = [encoder_hidden_states[:, 0].unsqueeze(1)]  # <BOS>
-            for i in range(1, args.max_token_length, tokenizer.model_max_length):
-                chunk = encoder_hidden_states[:, i : i + tokenizer.model_max_length - 2]  # <BOS> の後から 最後の前まで
-                if i > 0:
-                    for j in range(len(chunk)):
-                        if input_ids[j, 1] == tokenizer.eos_token:  # 空、つまり <BOS> <EOS> <PAD> ...のパターン
-                            chunk[j, 0] = chunk[j, 1]  # 次の <PAD> の値をコピーする
-                states_list.append(chunk)  # <BOS> の後から <EOS> の前まで
-            states_list.append(encoder_hidden_states[:, -1].unsqueeze(1))  # <EOS> か <PAD> のどちらか
-            encoder_hidden_states = torch.cat(states_list, dim=1)
-        else:
-            # v1: <BOS>...<EOS> の三連を <BOS>...<EOS> へ戻す
-            states_list = [encoder_hidden_states[:, 0].unsqueeze(1)]  # <BOS>
-            for i in range(1, args.max_token_length, tokenizer.model_max_length):
-                states_list.append(
-                    encoder_hidden_states[:, i : i + tokenizer.model_max_length - 2]
-                )  # <BOS> の後から <EOS> の前まで
-            states_list.append(encoder_hidden_states[:, -1].unsqueeze(1))  # <EOS>
-            encoder_hidden_states = torch.cat(states_list, dim=1)
-
-    if weight_dtype is not None:
-        # this is required for additional network training
-        encoder_hidden_states = encoder_hidden_states.to(weight_dtype)
-
-    return encoder_hidden_states
-
-
-def pool_workaround(
-    text_encoder: CLIPTextModelWithProjection, last_hidden_state: torch.Tensor, input_ids: torch.Tensor, eos_token_id: int
-):
-    r"""
-    workaround for CLIP's pooling bug: it returns the hidden states for the max token id as the pooled output
-    instead of the hidden states for the EOS token
-    If we use Textual Inversion, we need to use the hidden states for the EOS token as the pooled output
-
-    Original code from CLIP's pooling function:
-
-    \# text_embeds.shape = [batch_size, sequence_length, transformer.width]
-    \# take features from the eot embedding (eot_token is the highest number in each sequence)
-    \# casting to torch.int for onnx compatibility: argmax doesn't support int64 inputs with opset 14
-    pooled_output = last_hidden_state[
-        torch.arange(last_hidden_state.shape[0], device=last_hidden_state.device),
-        input_ids.to(dtype=torch.int, device=last_hidden_state.device).argmax(dim=-1),
-    ]
-    """
-
-    # input_ids: b*n,77
-    # find index for EOS token
-
-    # Following code is not working if one of the input_ids has multiple EOS tokens (very odd case)
-    # eos_token_index = torch.where(input_ids == eos_token_id)[1]
-    # eos_token_index = eos_token_index.to(device=last_hidden_state.device)
-
-    # Create a mask where the EOS tokens are
-    eos_token_mask = (input_ids == eos_token_id).int()
-
-    # Use argmax to find the last index of the EOS token for each element in the batch
-    eos_token_index = torch.argmax(eos_token_mask, dim=1)  # this will be 0 if there is no EOS token, it's fine
-    eos_token_index = eos_token_index.to(device=last_hidden_state.device)
-
-    # get hidden states for EOS token
-    pooled_output = last_hidden_state[torch.arange(last_hidden_state.shape[0], device=last_hidden_state.device), eos_token_index]
-
-    # apply projection: projection may be of different dtype than last_hidden_state
-    pooled_output = text_encoder.text_projection(pooled_output.to(text_encoder.text_projection.weight.dtype))
-    pooled_output = pooled_output.to(last_hidden_state.dtype)
-
-    return pooled_output
-
-
-def get_hidden_states_sdxl(
-    max_token_length: int,
-    input_ids1: torch.Tensor,
-    input_ids2: torch.Tensor,
-    tokenizer1: CLIPTokenizer,
-    tokenizer2: CLIPTokenizer,
-    text_encoder1: CLIPTextModel,
-    text_encoder2: CLIPTextModelWithProjection,
-    weight_dtype: Optional[str] = None,
-    accelerator: Optional[Accelerator] = None,
-):
-    # input_ids: b,n,77 -> b*n, 77
-    b_size = input_ids1.size()[0]
-    input_ids1 = input_ids1.reshape((-1, tokenizer1.model_max_length))  # batch_size*n, 77
-    input_ids2 = input_ids2.reshape((-1, tokenizer2.model_max_length))  # batch_size*n, 77
-
-    # text_encoder1
-    enc_out = text_encoder1(input_ids1, output_hidden_states=True, return_dict=True)
-    hidden_states1 = enc_out["hidden_states"][11]
-
-    # text_encoder2
-    enc_out = text_encoder2(input_ids2, output_hidden_states=True, return_dict=True)
-    hidden_states2 = enc_out["hidden_states"][-2]  # penuultimate layer
-
-    # pool2 = enc_out["text_embeds"]
-    unwrapped_text_encoder2 = text_encoder2 if accelerator is None else accelerator.unwrap_model(text_encoder2)
-    pool2 = pool_workaround(unwrapped_text_encoder2, enc_out["last_hidden_state"], input_ids2, tokenizer2.eos_token_id)
-
-    # b*n, 77, 768 or 1280 -> b, n*77, 768 or 1280
-    n_size = 1 if max_token_length is None else max_token_length // 75
-    hidden_states1 = hidden_states1.reshape((b_size, -1, hidden_states1.shape[-1]))
-    hidden_states2 = hidden_states2.reshape((b_size, -1, hidden_states2.shape[-1]))
-
-    if max_token_length is not None:
-        # bs*3, 77, 768 or 1024
-        # encoder1: <BOS>...<EOS> の三連を <BOS>...<EOS> へ戻す
-        states_list = [hidden_states1[:, 0].unsqueeze(1)]  # <BOS>
-        for i in range(1, max_token_length, tokenizer1.model_max_length):
-            states_list.append(hidden_states1[:, i : i + tokenizer1.model_max_length - 2])  # <BOS> の後から <EOS> の前まで
-        states_list.append(hidden_states1[:, -1].unsqueeze(1))  # <EOS>
-        hidden_states1 = torch.cat(states_list, dim=1)
-
-        # v2: <BOS>...<EOS> <PAD> ... の三連を <BOS>...<EOS> <PAD> ... へ戻す　正直この実装でいいのかわからん
-        states_list = [hidden_states2[:, 0].unsqueeze(1)]  # <BOS>
-        for i in range(1, max_token_length, tokenizer2.model_max_length):
-            chunk = hidden_states2[:, i : i + tokenizer2.model_max_length - 2]  # <BOS> の後から 最後の前まで
-            # this causes an error:
-            # RuntimeError: one of the variables needed for gradient computation has been modified by an inplace operation
-            # if i > 1:
-            #     for j in range(len(chunk)):  # batch_size
-            #         if input_ids2[n_index + j * n_size, 1] == tokenizer2.eos_token_id:  # 空、つまり <BOS> <EOS> <PAD> ...のパターン
-            #             chunk[j, 0] = chunk[j, 1]  # 次の <PAD> の値をコピーする
-            states_list.append(chunk)  # <BOS> の後から <EOS> の前まで
-        states_list.append(hidden_states2[:, -1].unsqueeze(1))  # <EOS> か <PAD> のどちらか
-        hidden_states2 = torch.cat(states_list, dim=1)
-
-        # pool はnの最初のものを使う
-        pool2 = pool2[::n_size]
-
-    if weight_dtype is not None:
-        # this is required for additional network training
-        hidden_states1 = hidden_states1.to(weight_dtype)
-        hidden_states2 = hidden_states2.to(weight_dtype)
-
-    return hidden_states1, hidden_states2, pool2
+# Text encoder hidden-state helpers have moved to library.hidden_states;
+# re-exported here for backward compatibility.
+# New code should import from library.hidden_states directly.
+from library.hidden_states import (  # noqa: F401
+    get_hidden_states,
+    pool_workaround,
+    get_hidden_states_sdxl,
+)
 
 
 def default_if_none(value, default):
